+++ mse-20190531/alibabacloud_mse20190531/client.py
@@ -4677,6 +4677,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.accept_language):
             query['AcceptLanguage'] = request.accept_language
+        if not UtilClient.is_unset(request.encrypt_tool_spec):
+            query['EncryptToolSpec'] = request.encrypt_tool_spec
         if not UtilClient.is_unset(request.instance_id):
             query['InstanceId'] = request.instance_id
         if not UtilClient.is_unset(request.namespace_id):
@@ -4728,6 +4730,8 @@ class Client(OpenApiClient):
         query = {}
         if not UtilClient.is_unset(request.accept_language):
             query['AcceptLanguage'] = request.accept_language
+        if not UtilClient.is_unset(request.encrypt_tool_spec):
+            query['EncryptToolSpec'] = request.encrypt_tool_spec
         if not UtilClient.is_unset(request.instance_id):
             query['InstanceId'] = request.instance_id
         if not UtilClient.is_unset(request.namespace_id):